ctrlos: Add unsupported attribute quasi-aliases
Those attributes keep existing even when `allowAliases` is false.

This is important to keep evaluation working when removing unsupported
attributes. The dependents may be relying on a versioned package, and
cleaning up *every* dependent, or worse, updating every dependent is a
chore that cannot withstand the test of time.

So instead, we kind of “hibernate” those packages by making bogus
attributes that fails their specific evaluation. If a customer ends-up
needing support for a now unsupported dependent, we can thaw it as
needed.

While this is not checked currently, the packages that are unsupported
themselves should be removed from the package set entirely.