ci: Init scheduled jobs framework
This seems a bit weirdly over-engineered, but this is coming from
previous experience with long YAML files full of repeated constructs.
It's much neater to read `extends:` lists that describes a few common
constructs, which are configured with `variables`, than to copy a large
boilerplate object with a large script into it.