intel-media-sdk: mark vulnerable
The upstream repo was archived on 2023-05-17 with the following comment:

> This project has been identified as having known security escapes.

It is now clear what (some) of these are, which prompts marking this
package as vulnerable.

This is now safe to do, as it is not a default part of our ffmpeg-full
package any longer.

(cherry picked from commit c7a9aa54ec7f1a14e4a243c22bb13084620e295a)